Adds prop3D wrapper
