+++ pyvista/_vtk.py
@@ -498,6 +498,7 @@ if VTK9:
         vtkPointPicker,
         vtkPolyDataMapper,
         vtkPolyDataMapper2D,
+        vtkProp3D,
         vtkPropAssembly,
         vtkProperty,
         vtkPropPicker,
