+++ pyvista/plotting/volume.py
@@ -3,9 +3,10 @@ import pyvista as pv
 
 from ._property import Property
 from .mapper import _BaseMapper
+from .prop3D import Prop3D
 
 
-class Volume(pv._vtk.vtkVolume):
+class Volume(Prop3D, pv._vtk.vtkVolume):
     """Wrapper class for VTK volume.
 
     This class represents a volume in a rendered scene.
